requirements.txt
fugle-marketdata==2.4.1
streamlit>=1.32.0
plotly>=5.20.0
pandas>=2.0.0
python-dotenv>=1.0.0
